Addition of the makefile
added the makefile for the Card Game. If using the make clean command on linux the make file needs to be changed under clean: to rm *.o "name of the program". If using the make clean command on windows it needs to be changed to del *.o "name of program with ".exe. where the name of the program is whate ever executable the make will compiling everything into.